requirements.txt
amqp==2.5.2
appnope==0.1.0
APScheduler==3.6.3
asyncio-cancel-token==0.1.0a2
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.0
billiard==3.6.3.0
bleach==3.1.4
blinker==1.4
bs4==0.0.1
cached-property==1.5.1
celery==4.4.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
chart-studio==1.1.0
click==7.1.1
colorlover==0.3.0
coverage==5.0.4
cufflinks==0.17.3
cvxopt==1.2.5
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
deprecation==2.0.7
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-APScheduler==1.11.0
Flask-Bcrypt==0.7.1
Flask-Compress==1.4.0
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
Flask-WTF-Ext==0.2
frozendict==1.2
future==0.18.2
gitdb==4.0.2
GitPython==3.1.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.7
importlib-metadata==1.6.0
infinity==1.4
inflection==0.3.1
intervals==0.8.1
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jsonschema==2.6.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
kombu==4.6.8
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.5
normalize==2.0.2
notebook==6.0.3
numpy==1.18.2
packaging==20.3
pandas==1.0.3
pandas-datareader==0.8.1
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
pip-chill==1.0.0
plotly==4.6.0
pluggy==0.7.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
psutil==5.7.0
psycopg2==2.8.4
ptyprocess==0.6.0
py==1.8.1
py4j==0.10.7
pycodestyle==2.5.0
pycparser==2.20
pyEX==0.2.1
pyflakes==2.1.1
Pygments==2.6.1
pyparsing==2.4.6
pyrsistent==0.16.0
pyspark==2.4.0
pytest==3.6.4
pytest-asyncio==0.10.0
pytest-asyncio-network-simulator==0.1.0a2
pytest-cov==2.8.1
pytest-timeout==1.3.4
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
PyYAML==3.13
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
redis==3.4.1
requests==2.20.1
retrying==1.3.3
richenum==1.3.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.14.0
smmap==3.0.1
socketIO-client-nexus==0.7.6
soupsieve==2.0
SQLAlchemy==1.3.15
SQLAlchemy-ORM-tree==0.1.2
SQLAlchemy-Utils==0.36.3
sseclient==0.0.26
temporal-cache==0.1.1
terminado==0.8.3
testpath==0.4.4
texttable==0.9.1
tornado==6.0.4
traitlets==4.3.3
tzlocal==2.0.0
unique==0.1
urllib3==1.24.3
validators==0.14.3
vine==1.3.0
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.1
widgetsnbextension==3.5.1
WTForms==2.2.1
WTForms-Alchemy==0.16.9
WTForms-Components==0.9.7
WTForms-Ext==0.2
zipp==3.1.0
